requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Pillow==6.0.0
djangorestframework==3.9.0
django-filter==2.0.0
mock==2.0.0
